add tests to presenter
